Use signed type for counter variables of parallelized for loops.

+++ cpp/subprojects/boosting/src/boosting/output/predictor_classification_example_wise.cpp
@@ -92,7 +92,7 @@ namespace boosting {
         #pragma omp parallel for firstprivate(numExamples) firstprivate(numLabels) firstprivate(scoreMatrixPtr) \
         firstprivate(predictionMatrixPtr) firstprivate(measurePtr) firstprivate(labelVectors) schedule(dynamic) \
         num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             const LabelVector* closestLabelVector = findClosestLabelVector(scoreMatrixPtr->row_cbegin(i),
                                                                            scoreMatrixPtr->row_cend(i), *measurePtr,
                                                                            labelVectors);
@@ -113,7 +113,7 @@ namespace boosting {
         #pragma omp parallel for firstprivate(numExamples) firstprivate(numLabels) firstprivate(modelPtr) \
         firstprivate(featureMatrixPtr) firstprivate(predictionMatrixPtr) firstprivate(measurePtr) \
         firstprivate(labelVectors) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float64 scoreVector[numLabels] = {};
             applyRules(*modelPtr, featureMatrixPtr->row_cbegin(i), featureMatrixPtr->row_cend(i), &scoreVector[0]);
             const LabelVector* closestLabelVector = findClosestLabelVector(&scoreVector[0], &scoreVector[numLabels],
@@ -136,7 +136,7 @@ namespace boosting {
         #pragma omp parallel for firstprivate(numExamples) firstprivate(numFeatures) firstprivate(numLabels) \
         firstprivate(modelPtr) firstprivate(featureMatrixPtr) firstprivate(predictionMatrixPtr) \
         firstprivate(measurePtr) firstprivate(labelVectors) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float64 scoreVector[numLabels] = {};
             applyRulesCsr(*modelPtr, numFeatures, featureMatrixPtr->row_indices_cbegin(i),
                           featureMatrixPtr->row_indices_cend(i), featureMatrixPtr->row_values_cbegin(i),
@@ -161,7 +161,7 @@ namespace boosting {
         #pragma omp parallel for reduction(+:numNonZeroElements) firstprivate(numExamples) firstprivate(numLabels) \
         firstprivate(modelPtr) firstprivate(featureMatrixPtr) firstprivate(predictionMatrixPtr) \
         firstprivate(measurePtr) firstprivate(labelVectors) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float64 scoreVector[numLabels] = {};
             applyRules(*modelPtr, featureMatrixPtr->row_cbegin(i), featureMatrixPtr->row_cend(i), &scoreVector[0]);
             const LabelVector* closestLabelVector = findClosestLabelVector(&scoreVector[0], &scoreVector[numLabels],
@@ -188,7 +188,7 @@ namespace boosting {
         firstprivate(numLabels) firstprivate(modelPtr) firstprivate(featureMatrixPtr) \
         firstprivate(predictionMatrixPtr) firstprivate(measurePtr) firstprivate(labelVectors) schedule(dynamic) \
         num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float64 scoreVector[numLabels] = {};
             applyRulesCsr(*modelPtr, numFeatures, featureMatrixPtr->row_indices_cbegin(i),
                           featureMatrixPtr->row_indices_cend(i), featureMatrixPtr->row_values_cbegin(i),
